UI Overlay finalized.
Now works! Does not look great, but the underlying logic should be solid
(although it will be very interesting to see how it handles changes in
the data. Should be fine?)

Also added a more contrived flag system, which makes sure that flags
remain checked through a whole iteration. A bit verbose, but works well,
and should be more maintainable.

Next step is to add time-filtering.

I am going to merge this with Master, since the uiOverlay basics are
finalized.